+++ src/ui/layout/Row.tsx
@@ -7,6 +7,8 @@ interface Props {
   children: React.ReactNode
   align?: ItemsAlign
   gap: number
+  flex?: number
+  onClick?: () => void
 }
 
 const defaultProps: Props = {
@@ -14,9 +16,16 @@ const defaultProps: Props = {
   gap: 0
 }
 
-function Row({ children, align, gap }: Props) {
+function Row({ children, align, gap, flex, onClick }: Props) {
   return (
-    <Container align={align} gap={gap}>{children}</Container>
+    <Container
+      align={align}
+      gap={gap}
+      flex={flex}
+      onClick={onClick}
+    >
+      {children}
+    </Container>
   )
 }
 
@@ -25,6 +34,7 @@ const Container = styled.div<Partial<Props>>`
   align-items: center;
 
   ${({ gap }) => gap ? `gap: ${gap}px;` : ''}
+  ${({ flex }) => flex ? `flex: ${flex};` : ''}
 
   ${({ align }) => {
     let styles: string = ''