+++ src/ui/organisms/OrderItem/ImageContent.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import styled from '@emotion/styled'
 
 import Icon from '../../atoms/Icon'
+import Image from '../../atoms/Image'
 
 interface ImageContentProps {
   image?: string
@@ -11,29 +12,17 @@ function ImageContent({ image }: ImageContentProps) {
   return (
     <ImageContainer>
       {!image ? (
-        <Icon name='image' size={35} />
+        <Icon name='image' size={70} withBackground={0.5} />
       ) : (
-        <Image src={image} />
+        <Image src={image} width={70} height={70} />
       )}
     </ImageContainer>
   )
 }
 
 const ImageContainer = styled.figure`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 4.4rem;
-  height: 4.4rem;
-  background-color: ${({ theme }) => theme.colors.NEUTRAL.SIDE};
-  border-radius: 10px;
-  padding: 0.5rem;
-`
-
-const Image = styled.img`
-  width: 100%;
-  height: 100%;
-  object-fit: contain;
+  width: 70px;
+  height: 70px;
 `
 
 export default ImageContent