+++ src/ui/organisms/OrderSnapshot/OrderHeader.tsx
@@ -19,18 +19,23 @@ type Props = {
   orderId: string | number
   date: string
   status: string
+  showStatusChip: boolean,
 }
 
-function OrderHeader({ orderId, status, date }: Props) {
+function OrderHeader({ orderId, status, showStatusChip, date }: Props) {
   const { colors } = useTheme()
 
   return (
     <Row align='space-between'>
-      <StatusChip
-        type={statusAsTypes[status]}
-        text={String(orderId)}
-        textColor={colors.FONT.TITLE}
-      />
+      {showStatusChip ? (
+        <StatusChip
+          type={statusAsTypes[status]}
+          text={String(orderId)}
+          textColor={colors.FONT.TITLE}
+        />
+      ) : (
+        <Typography weight={700} size={16}>{String(orderId)}</Typography>
+      )}
 
       <Typography type='description' size={14}>{dateFormat(date, 'es-CL', true)}</Typography>
     </Row>